requirements.txt
telethon
pillow
pytesseract
